Split tests and stories for Form and useKeyboard
